# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Relationship CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List organizations
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationListResponse"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
...
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: organizationId
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactListResponse"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "409":
          description: Duplicate email
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          schema:
            type: string
        - name: contactId
          in: query
          schema:
            type: string
        - name: organizationId
          in: query
          schema:
            type: string
        - name: dueBefore
          in: query
          schema:
            type: string
        - name: dueAfter
          in: query
          schema:
            type: string
        - name: mine
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /activities:
    get:
      operationId: listActivities
      tags: [activities]
      summary: List activities
      parameters:
        - name: contactId
          in: query
          schema:
            type: string
        - name: organizationId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of activities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
    post:
      operationId: createActivity
      tags: [activities]
      summary: Create activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActivityBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        status:
          type: string
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        ownerId:
          type: string
        ownerName:
          type: ["string", "null"]
        contactCount:
          type: integer
        lastActivityAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - status
        - ownerId
        - contactCount
        - createdAt
        - updatedAt
    OrganizationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages
...
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        status:
          type: string
        location:
          type: string
        notes:
          type: string
        ownerId:
          type: string
      required:
        - name
        - type
...
    Contact:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        status:
          type: string
        lastContactedAt:
          type: ["string", "null"]
          format: date-time
        organizationId:
          type: string
        organizationName:
          type: ["string", "null"]
        ownerId:
          type: string
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - status
        - organizationId
        - ownerId
        - createdAt
        - updatedAt
    ContactListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - data
        - total
        - page
        - totalPages
...
    CreateContactBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        status:
          type: string
        organizationId:
          type: string
        ownerId:
          type: string
      required:
        - firstName
        - lastName
        - email
        - organizationId
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: string
          format: date-time
        priority:
          type: string
        status:
          type: string
        ownerId:
          type: string
        ownerName:
          type: ["string", "null"]
        contactId:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        organizationId:
          type: ["string", "null"]
        organizationName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - dueDate
        - priority
        - status
        - ownerId
        - createdAt
        - updatedAt
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        priority:
          type: string
        status:
          type: string
        ownerId:
          type: string
        contactId:
          type: string
        organizationId:
          type: string
      required:
        - title
        - dueDate
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        priority:
          type: string
        status:
          type: string
        ownerId:
          type: string
        contactId:
          type: string
        organizationId:
          type: string
    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        summary:
          type: string
        date:
          type: string
          format: date-time
        contactId:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        organizationId:
          type: ["string", "null"]
        organizationName:
          type: ["string", "null"]
        userId:
          type: string
        userName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - summary
        - date
        - userId
        - createdAt
    CreateActivityBody:
      type: object
      properties:
        type:
          type: string
        summary:
          type: string
        date:
          type: string
          format: date-time
        contactId:
          type: string
        organizationId:
          type: string
      required:
        - type
        - summary
...